+++ src/backend/joanie/tests/core/test_api_organizations_contract.py
@@ -175,6 +175,7 @@ class OrganizationContractApiTest(BaseAPITestCase):
                                 "id": str(contract.order.course.id),
                                 "title": contract.order.course.title,
                             },
+                            "enrollment": None,
                             "organization": {
                                 "id": str(contract.order.organization.id),
                                 "code": contract.order.organization.code,
@@ -401,6 +402,7 @@ class OrganizationContractApiTest(BaseAPITestCase):
                         "id": str(contract.order.course.id),
                         "title": contract.order.course.title,
                     },
+                    "enrollment": None,
                     "organization": {
                         "id": str(contract.order.organization.id),
                         "code": contract.order.organization.code,